Concurrency
As we are not shocked to discover, sorting through the AnimatedTurtle teleport bug #1 is an exercise in concurrency. It surfaced some bad behavior on my part, since cleaned up, in terms of instantiating the app frame. Empirically, things seem to be working, but there's a FIXME on `getX()` that I want to reason with before committing to being done with this problem.